Don't use src and utility folders
The Arduino library 1.5 format does not allow the use of both src and utility folders in the root of the library. If a library has both of these folders then if will fail to compile:

Library can't use both 'src' and 'utility' folders.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#source-code
