+++ extras/EEPROM/examples/eeprom_clear/eeprom_clear.ino
