+++ extras/RFRemocon/keywords.txt
@@ -1,0 +1,37 @@
+#######################################
+# Syntax Coloring Map For RFRemocon
+#######################################
+
+#######################################
+# Datatypes (KEYWORD1)
+#######################################
+
+RFRemocon	KEYWORD1
+
+
+#######################################
+# Methods and Functions (KEYWORD2)
+#######################################
+
+attachSerial	KEYWORD2
+begin	KEYWORD2
+flush	KEYWORD2
+send	KEYWORD2
+recieve	KEYWORD2
+
+
+#######################################
+# Constants (LITERAL1)
+#######################################
+
+KEY_UP	LITERAL1
+KEY_U	LITERAL1
+KEY_D	LITERAL1
+KEY_L	LITERAL1
+KEY_R	LITERAL1
+KEY_1	LITERAL1
+KEY_2	LITERAL1
+KEY_3	LITERAL1
+KEY_4	LITERAL1
+KEY_5	LITERAL1
+KEY_6	LITERAL1
