--- utility/Servo/keywords.txt
@@ -1,24 +1,0 @@
-#######################################
-# Syntax Coloring Map Servo
-#######################################
-
-#######################################
-# Datatypes (KEYWORD1)
-#######################################
-
-Servo	KEYWORD1
-
-#######################################
-# Methods and Functions (KEYWORD2)
-#######################################
-attach	KEYWORD2
-detach	KEYWORD2
-write	KEYWORD2
-read	KEYWORD2
-attached	KEYWORD2
-writeMicroseconds	KEYWORD2
-readMicroseconds	KEYWORD2
-
-#######################################
-# Constants (LITERAL1)
-#######################################
